requirements.txt
streamlit
supabase==2.10.0
python-dotenv